Refactor regex parsing into separate file and create endpoint to process all the JSON files from a given folder
